Fix bug: dangerous comparison of $thread['closed']
The `closed` column of the `threads` table defaults to an empty string,
which in PHP does *not* compare equal to zero. Thus, when testing for an
open thread, we need to test for `closed` *not equal to one* rather than
`closed` *equal to zero*. (An additional factor to consider is that this
column is a varchar which can be set to `moved|...` to indicate a moved
thread).